Moved Java Project To Maven
I have moved the Java Project to Maven which will allow JUnit tests to be completed on code to ensure all changes work before updating to master branch.
