--- EncryptionManager/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=9fcfd159
-build.xml.script.CRC32=5d767b9e
-build.xml.stylesheet.CRC32=f85dc8f2@1.89.1.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=9fcfd159
-nbproject/build-impl.xml.script.CRC32=8cd01ee1
-nbproject/build-impl.xml.stylesheet.CRC32=3a2fa800@1.89.1.48
