+++ src/main/java/com/encryptionmanager/DatabaseManager.java
@@ -1,4 +1,6 @@
-package encryptionmanager;
+package com.encryptionmanager;
+
+
 
 import java.sql.Connection;
 import java.sql.DriverManager;
