+++ src/main/java/com/encryptionmanager/EncryptionMethods.java
@@ -1,4 +1,6 @@
-package encryptionmanager;
+package com.encryptionmanager;
+
+
 
 import java.security.InvalidAlgorithmParameterException;
 import java.security.SecureRandom;
