+++ src/main/java/com/encryptionmanager/ValidationMethods.java
@@ -1,4 +1,6 @@
-package encryptionmanager;
+package com.encryptionmanager;
+
+
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
